// src/@types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";


export interface IFile {
  url: string;
  details: {
    size: number;
    image: {
      width: number;
      height: number;
    };
  };
  fileName: string;
  contentType: string;
}

export interface IAuthorFields {
  /** Name */
  name: string;
  /** Picture */
  picture: Asset;
}

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "author";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ICategoryFields {
  /** Name */
  name: string;
  /** Slug */
  slug: string;
}

export interface ICategory extends Entry<ICategoryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "category";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPostFields {
  /** Title */
  title: string;
  /** Slug */
  slug: string;
  /** Author */
  author: IAuthor;
  /** Excerpt */
  excerpt: string;
  /** Cover Image */
  coverImage: Asset;
  /** Content */
  content: Document;
  /** Date */
  date: string;
  /** Category */
  category: ICategory[];
}

export interface IPost extends Entry<IPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "post";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}



export type CONTENT_TYPE = "author" | "category" | "post";
export type IEntry = IAuthor | ICategory | IPost;
export type LOCALE_CODE = "en-US";
export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";